# internal/app/model/property.go
package model

import (
	"time"

	"gorm.io/gorm"
)

type Property struct {
	ID          uint           `gorm:"primaryKey" json:"id"`
	CreatedAt   time.Time      `json:"created_at"`
	UpdatedAt   time.Time      `json:"updated_at"`
	DeletedAt   gorm.DeletedAt `gorm:"index" json:"deleted_at"`
	Name        string         `json:"name"`
	Description string         `json:"description"`
	Price       float64        `json:"price"`
	Location    string         `json:"location"`
	Images      []Image        `gorm:"foreignKey:PropertyID" json:"images"`
}

type Image struct {
	ID         uint           `gorm:"primaryKey" json:"id"`
	CreatedAt  time.Time      `json:"created_at"`
	UpdatedAt  time.Time      `json:"updated_at"`
	DeletedAt  gorm.DeletedAt `gorm:"index" json:"deleted_at"`
	PropertyID uint           `json:"property_id"`
	URL        string         `json:"url"`
}